+++ app/service/health_check.py
@@ -1,33 +1,27 @@
 from loguru import logger
-
 from app.config import get_settings
 from app.db import engine
-from app.storage.aws_s3 import s3_client
+from app.storage.s3 import s3_client
 
 settings = get_settings()
 
-
-def test_db():
+def test_db() -> dict[str, str]:
     try:
         with engine.connect():
             return {"db": "healthy"}
     except Exception as err:
-        logger.exception(err)
+        logger.exception("Database connection failed: {}", err)
         raise err
 
-
-def test_storage():
+def test_storage() -> dict[str, str]:
     try:
-        s3_client.head_bucket(Bucket=settings.s3_bucket_name)  # response =
-        # print("@@@@@@@@@@@@@", response)
+        s3_client.head_bucket(Bucket=settings.s3_bucket_name)
+        return {"storage": "healthy"}
     except Exception as err:
-        logger.exception(err)  # , exc_info=True
+        logger.exception("S3 storage check failed: {}", err)
         raise err
-
-    return {"storage": "healthy"}
-
 
 def run_healthcheck() -> dict[str, str]:
-    test_db()
-    test_storage()
-    return {"status": "ALIVE"}
+    db_status = test_db()
+    storage_status = test_storage()
+    return {"status": "ALIVE", **db_status, **storage_status}
