--- app/storage/aws_s3.py
@@ -1,25 +1,0 @@
-import boto3
-from app.config import get_settings
-
-settings = get_settings()
-
-s3_client = boto3.client(
-    "s3",
-    region_name=settings.s3_region,
-    aws_access_key_id=settings.s3_access_key,
-    aws_secret_access_key=settings.s3_secret_access_key,
-)
-
-s3_resource = boto3.resource(
-    "s3",
-    region_name=settings.s3_region,
-    aws_access_key_id=settings.s3_access_key,
-    aws_secret_access_key=settings.s3_secret_access_key,
-)
-
-def generate_presigned_url(bucket: str, key: str, expiration: int = 3600) -> str:
-    return s3_client.generate_presigned_url(
-        "get_object",
-        Params={"Bucket": bucket, "Key": key},
-        ExpiresIn=expiration,
-    )
